requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pandas==2.2.2
scikit-learn==1.4.2
xgboost==2.0.3
pydantic==2.7.1
openpyxl==3.1.2
Pillow==10.3.0